🧹 use the internal module pattern for @atjson/document
When testing adding the Schema definition in a test branch, this caused dependency loops with imports. This pattern fixes dependency loops.
