+++ packages/@atjson/document/src/annotations/inline.ts
@@ -1,4 +1,4 @@
-import Annotation from "../annotation";
+import { Annotation } from "../internals";
 
 export abstract class InlineAnnotation<Attributes = {}> extends Annotation<
   Attributes
