+++ packages/@atjson/document/src/json.ts
@@ -6,5 +6,3 @@ export interface Dictionary<T> {
 export type JSON = string | number | boolean | null | JSONObject | JSONArray;
 export interface JSONObject extends Dictionary<JSON> {}
 export interface JSONArray extends Array<JSON> {}
-
-export default JSON;
